moved server info out into it's own file
